+++ doc_swift_ui_gallery/swift_ui_gallery/swift_ui_gallery/shapes/RectangleExamples.swift
@@ -23,7 +23,8 @@ struct RectanglePage: View {
                 Rectangle()
                     .stroke(LinearGradient(gradient: Gradient(colors: [.yellow, .blue]), startPoint: .leading, endPoint: .trailing), lineWidth: 14)
                     .frame(width: 200, height: 100)
-            }
+                
+            }.frame(width: 300)
         }
     }
 }