+++ test/shapes/shadow_golden_test.dart
@@ -1,58 +1,59 @@
-import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:golden_toolkit/golden_toolkit.dart';
 import 'package:swift_ui/swift_ui.dart';
 
 void main() {
-  group("Shapes > Rectangle >", () {
+  group("Shapes > Shadow >", () {
     testGoldens("smoke test", (widgetTester) async {
       final builder = GoldenBuilder.grid(columns: 2, widthToHeightRatio: 1)
         ..addScenario(
-          'Solid fill',
+          'Rectangle',
           const Frame(
             width: 200,
             height: 100,
-            Rectangle(
-              fillColor: Colors.blue,
-            ),
-          ),
-        )
-        ..addScenario(
-          'Gradient fill',
-          const Frame(
-            width: 200,
-            height: 100,
-            Rectangle(
-              fillGradient: LinearGradient(
-                colors: [Colors.yellow, Colors.orange],
+            Shadow(
+              radius: 5,
+              x: 5,
+              y: 5,
+              Rectangle(
+                fillColor: Colors.blue,
               ),
             ),
           ),
         )
         ..addScenario(
-          'Solid stroke',
+          'Ellipse',
           const Frame(
             width: 200,
             height: 100,
-            Rectangle(
-              strokeColor: Colors.blue,
+            Shadow(
+              radius: 5,
+              x: 5,
+              y: 5,
+              Ellipse(
+                fillColor: Colors.blue,
+              ),
             ),
           ),
         )
         ..addScenario(
-          'Gradient stroke',
+          'Ellipse stroke',
           const Frame(
             width: 200,
             height: 100,
-            Rectangle(
-              strokeGradient: LinearGradient(
-                colors: [Colors.yellow, Colors.orange],
+            Shadow(
+              radius: 5,
+              x: 5,
+              y: 5,
+              Ellipse(
+                strokeLineWidth: 5,
+                strokeColor: Colors.blue,
               ),
             ),
           ),
         );
       await widgetTester.pumpWidgetBuilder(builder.build());
-      await screenMatchesGolden(widgetTester, 'rectangle_smoke-test');
+      await screenMatchesGolden(widgetTester, 'shadow_smoke-test');
     });
   });
 }