+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Added `HStack` widget.
 - Added `VStack` widget.
 - Added `Rectangle` widget.
+- Added `Ellipse` widget.
+- Added `Shadow` widget.
 
 ## 0.0.1 - Sept, 2023
 Setting up the project structure. This release is not usable.